Move time utils under utils/times.py
